add: slices/offers.ts
